# Conflicts:
#	package-lock.json
#	package.json
#	src/NetscriptJSEvaluator.ts
#	src/NetscriptWorker.ts
#	src/Terminal/commands/download.ts
#	src/engine.tsx
#	src/ui/GameRoot.tsx
#	src/utils/JSONReviver.ts
#	test/jest/Netscript/RunScript.test.ts

+++ src/Augmentation/AugmentationHelpers.ts
@@ -12,12 +12,28 @@ import { Router } from "../ui/GameRoot";
 import { Page } from "../ui/Router";
 import { mergeMultipliers } from "../PersonObjects/Multipliers";
 import { currentNodeMults } from "../BitNode/BitNodeMultipliers";
+import { prestigeWorkerScripts } from "../NetscriptWorker";
+
+const soaAugmentationNames = [
+  AugmentationName.BeautyOfAphrodite,
+  AugmentationName.ChaosOfDionysus,
+  AugmentationName.FloodOfPoseidon,
+  AugmentationName.HuntOfArtemis,
+  AugmentationName.KnowledgeOfApollo,
+  AugmentationName.MightOfAres,
+  AugmentationName.TrickeryOfHermes,
+  AugmentationName.WKSharmonizer,
+  AugmentationName.WisdomOfAthena,
+];
 
 export function getBaseAugmentationPriceMultiplier(): number {
-  return CONSTANTS.MultipleAugMultiplier * [1, 0.96, 0.94, 0.93][Player.sourceFileLvl(11)];
+  return CONSTANTS.MultipleAugMultiplier * [1, 0.96, 0.94, 0.93][Player.activeSourceFileLvl(11)];
 }
 export function getGenericAugmentationPriceMultiplier(): number {
-  return Math.pow(getBaseAugmentationPriceMultiplier(), Player.queuedAugmentations.length);
+  const queuedNonSoAAugmentationList = Player.queuedAugmentations.filter((augmentation) => {
+    return !soaAugmentationNames.includes(augmentation.name);
+  });
+  return Math.pow(getBaseAugmentationPriceMultiplier(), queuedNonSoAAugmentationList.length);
 }
 
 export function applyAugmentation(aug: PlayerOwnedAugmentation, reapply = false): void {
@@ -52,6 +68,10 @@ export function installAugmentations(force?: boolean): boolean {
     dialogBoxCreate("You have not purchased any Augmentations to install!");
     return false;
   }
+
+  // We must kill all scripts before installing augmentations.
+  prestigeWorkerScripts();
+
   let augmentationList = "";
   let nfgIndex = -1;
   for (let i = Player.queuedAugmentations.length - 1; i >= 0; i--) {
@@ -78,7 +98,7 @@ export function installAugmentations(force?: boolean): boolean {
     augmentationList += aug.name + level + "\n";
   }
   Player.queuedAugmentations = [];
-  if (!force) {
+  if (!force && augmentationList !== "") {
     dialogBoxCreate(
       "You slowly drift to sleep as scientists put you under in order " +
         "to install the following Augmentations:\n" +
@@ -111,7 +131,7 @@ export function getAugCost(aug: Augmentation): AugmentationCosts {
       const multiplier = Math.pow(CONSTANTS.NeuroFluxGovernorLevelMult, aug.getLevel());
       repCost = aug.baseRepRequirement * multiplier * currentNodeMults.AugmentationRepCost;
       moneyCost = aug.baseCost * multiplier * currentNodeMults.AugmentationMoneyCost;
-      moneyCost *= getBaseAugmentationPriceMultiplier() ** Player.queuedAugmentations.length;
+      moneyCost *= getGenericAugmentationPriceMultiplier();
       break;
     }
     // SOA Augments use a unique cost method
@@ -124,17 +144,6 @@ export function getAugCost(aug: Augmentation): AugmentationCosts {
     case AugmentationName.TrickeryOfHermes:
     case AugmentationName.WKSharmonizer:
     case AugmentationName.WisdomOfAthena: {
-      const soaAugmentationNames = [
-        AugmentationName.BeautyOfAphrodite,
-        AugmentationName.ChaosOfDionysus,
-        AugmentationName.FloodOfPoseidon,
-        AugmentationName.HuntOfArtemis,
-        AugmentationName.KnowledgeOfApollo,
-        AugmentationName.MightOfAres,
-        AugmentationName.TrickeryOfHermes,
-        AugmentationName.WKSharmonizer,
-        AugmentationName.WisdomOfAthena,
-      ];
       const soaAugCount = soaAugmentationNames.filter((augName) => Player.hasAugmentation(augName)).length;
       moneyCost = aug.baseCost * Math.pow(CONSTANTS.SoACostMult, soaAugCount);
       repCost = aug.baseRepRequirement * Math.pow(CONSTANTS.SoARepMult, soaAugCount);
